+++ CHANGELOG.md
@@ -6,6 +6,19 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.11.1 - Unreleased
 
+### Added
+* `Points.to_shapely` and `Coords.to_shapely`
+* `Points.from_shapely` and `Coords.from_shapely`
+* `Points.wkt` and `Coords.wkt`
+* classes arg to `Points.coerce`
+
+
+### Changed
+* `kwcoco.CategoryTree` now returns other metadata with it when converting to COCO dictionaries 
+
+### Fixed
+* Issue in Points.coerce where unspecified keypoints categories were interpreted as specified with length zero.
+
 
 ## Version 0.11.0 - Released 2024-10-15